refactor: bottom sheet hook 구현과 maxWidth 수정
