+++ src/BottomSheet/useBottomSheet.ts
@@ -1,0 +1,26 @@
+import { useCallback, useRef, useState } from 'react';
+
+export const useBottomSheet = () => {
+  const [isClosing, setIsClosing] = useState(false);
+  const ref = useRef<HTMLDialogElement>(null);
+
+  const closeAnimated = useCallback(() => {
+    setIsClosing(true);
+
+    const timer = setTimeout(() => {
+      setIsClosing(false);
+      ref.current?.close();
+      clearTimeout(timer);
+    }, 200);
+  }, []);
+
+  const handleOpenBottomSheet = () => {
+    ref.current?.showModal();
+  };
+
+  const handleCloseBottomSheet = () => {
+    closeAnimated();
+  };
+
+  return { ref, isClosing, handleOpenBottomSheet, handleCloseBottomSheet };
+};
